# prefetch hint (1.3x speed-up)