--- a/include/LightGBM/meta.h
+++ b/include/LightGBM/meta.h
@@ -12,6 +12,13 @@
 #include <utility>
 #include <vector>
 
+#if (defined(_MSC_VER) && (defined(_M_IX86) || defined(_M_AMD64))) || defined(__INTEL_COMPILER)
+#include <xmmintrin.h>
+#define PREFETCH_T0(addr) _mm_prefetch(reinterpret_cast<const char*>(addr), _MM_HINT_T0)
+#elif defined(__GNUC__)
+#define PREFETCH_T0(addr) __builtin_prefetch(reinterpret_cast<const char*>(addr), 0, 3)
+#endif
+
 namespace LightGBM {
 
 /*! \brief Type of data size, it is better to use signed type*/