set prefetch hint to increase cache hit chance.

benchmark:

| Data      | Master|  This PR | Improvement |  
|----|  ----| ----- | ----|
| Higgs|252.526271  |192.186991 s | 1.31x |
| Yahoo LTR|158.277699 s |129.526579 s | 1.22x |
| MS LTR|357.12529 s|251.736296 s | 1.42x |