@Laurae2 2020-01-13
Here is a scalability (by number of threads) benchmark result on Bosch using non-sparse columns:

![image](https://user-images.githubusercontent.com/9083669/72290556-c7c24200-364d-11ea-9cf2-44223aeb50ca.png)

Hardware:

* 2x Dual Xeon 6154 (2x 18 cores / 36 threads)
* 768GB RAM

As we can notice, there is a substantial improvement below 36 threads (1 NUMA node), but using 2 NUMA nodes hit the performance.

Most number of threads were run 7 times but the tasks were extremely short (few seconds), hence the large variation.