@guolinke 2020-01-15
@Laurae2 I guess that `prefetch_size` may need to tune according to `num_threads`.